+++ playground/src/example/default.vue
@@ -1,0 +1,9 @@
+<template>
+  <ShadcnAlert>
+    <p>Welcome to the View Shadcn UI Playground</p>
+    <p>欢迎来到 View Shadcn UI 演练场</p>
+  </ShadcnAlert>
+</template>
+
+<script setup lang="ts">
+</script>