+++ vite.config.ts
@@ -1,13 +1,11 @@
 import path from 'path'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-
 import tailwind from 'tailwindcss'
 import autoprefixer from 'autoprefixer'
 import monacoEditorPlugin from 'vite-plugin-monaco-editor'
 
 export default defineConfig({
-    // base: '/devlive-community/shadcn-ui-vue-admin',
     css: {
         postcss: {
             plugins: [tailwind(), autoprefixer()]
@@ -15,7 +13,31 @@ export default defineConfig({
     },
     plugins: [
         vue(),
-        monacoEditorPlugin({})
+        monacoEditorPlugin({
+            customWorkers: [
+                {
+                    label: 'json',
+                    entry: 'monaco-editor/esm/vs/language/json/json.worker'
+                },
+                {
+                    label: 'css',
+                    entry: 'monaco-editor/esm/vs/language/css/css.worker'
+                },
+                {
+                    label: 'html',
+                    entry: 'monaco-editor/esm/vs/language/html/html.worker'
+                },
+                {
+                    label: 'typescript',
+                    entry: 'monaco-editor/esm/vs/language/typescript/ts.worker'
+                },
+                {
+                    label: 'editor',
+                    entry: 'monaco-editor/esm/vs/editor/editor.worker'
+                }
+            ],
+            publicPath: 'monaco-editor/min/worker'
+        })
     ],
     resolve: {
         alias: {
@@ -30,10 +52,11 @@ export default defineConfig({
             formats: ['es', 'umd']
         },
         rollupOptions: {
-            external: ['vue', '../dist/view-shadcn-ui.min.css'],
+            external: ['vue', '../dist/view-shadcn-ui.min.css', 'monaco-editor'],
             output: {
                 globals: {
-                    vue: 'Vue'
+                    vue: 'Vue',
+                    'monaco-editor': 'monaco'
                 }
             }
         }